feat: enable customized loss function in SAITS;
